fix ChangeLog for the unexpectedly committed changes. disable test_array_insert1 only on rubinius.
git-svn-id: svn+ssh://rubyforge.org/var/svn/ruby-oci8/branches/ruby-oci8-2.0@413 f081da96-e817-0410-9439-ba585a54bbf1

+++ test/test_array_dml.rb
@@ -13,7 +13,9 @@ class TestArrayDML < Test::Unit::TestCase
 
   # test inserting arrays with different data types
   #   including char, varchar2, number, date and so on
-  def _test_array_insert1
+  def test_array_insert1
+    # skip this test until the rubinius issue #445 is fixed.
+    return if defined? RUBY_ENGINE and RUBY_ENGINE == 'rbx'
     drop_table('test_table')
     sql = <<-EOS
 CREATE TABLE test_table
